requirements.txt
requests>=2.28.0
pyperclip>=1.8.2
python-dotenv>=0.20.0 